feat(App.tsx): add functionality to generate reverse accepted cards
The function handleGenerateReverseCards was added to send a POST request to the server with the accepted cards in markdown format. The server will generate the reverse cards and return them in JSON format. The reverse cards are then added to the existing cards array. If the request fails, an error message is displayed. A button was added to the UI to trigger this functionality.